+++ src/pat/inject/inject.js
@@ -661,30 +661,57 @@ const inject = {
         $el.off("pat-ajax-error.pat-inject");
     },
 
-    _onInjectError($el, cfgs, event) {
+    async _onInjectError($el, cfgs, event) {
         let explanation = "";
+        let fallback;
+        const status = event.jqxhr.status;
         const timestamp = new Date();
-        if (event.jqxhr.status % 100 == 4) {
+        if (status % 100 == 4) {
             explanation =
                 "Sorry! We couldn't find the page to load. Please make a screenshot and send it to support. Thank you!";
-        } else if (event.jqxhr.status % 100 == 5) {
+        } else if (status % 100 == 5) {
             explanation =
                 "I am very sorry! There was an error at the server. Please make a screenshot and contact support. Thank you!";
-        } else if (event.jqxhr.status == 0) {
+        } else if (status == 0) {
             explanation =
                 "It seems, the server is down. Please make a screenshot and contact support. Thank you!";
         }
-        const msg_attr = `${explanation} Status is ${event.jqxhr.status} ${event.jqxhr.statusText}, time was ${timestamp}. You can click to close this.`;
-        $("body").attr("data-error-message", msg_attr);
-        $("body").on("click", () => {
-            $("body").removeAttr("data-error-message");
-            window.location.href = window.location.href;
-        });
+
+        const fallback_url = document
+            .querySelector(`meta[name=pat-inject-${status}]`)
+            ?.getAttribute("content", false);
+        if (fallback_url) {
+            try {
+                const fallback_response = await fetch(fallback_url, {
+                    method: "GET",
+                });
+                fallback = document.createElement("html");
+                fallback.innerHTML = await fallback_response.text();
+                fallback = fallback.querySelector("body");
+            } catch {
+                // fallback to standard error message and ignore.
+            }
+        }
+
+        // clean up
         cfgs.forEach((cfg) => {
             if ("$injected" in cfg) cfg.$injected.remove();
         });
         $el.off("pat-ajax-success.pat-inject");
         $el.off("pat-ajax-error.pat-inject");
+
+        if (fallback) {
+            document.body.innerHTML = fallback.innerHTML;
+        } else {
+            const msg_attr =
+                fallback ||
+                `${explanation} Status is ${status} ${event.jqxhr.statusText}, time was ${timestamp}. You can click to close this.`;
+            $("body").attr("data-error-message", msg_attr);
+            $("body").on("click", () => {
+                $("body").removeAttr("data-error-message");
+                window.location.href = window.location.href; // reload
+            });
+        }
     },
 
     execute(cfgs, $el) {